+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * [#120](https://github.com/rubocop-hq/rubocop-rails/issues/120): Fix message for `Rails/SaveBang` when the save is in the body of a conditional. ([@jas14][])
 * [#131](https://github.com/rubocop-hq/rubocop-rails/pull/131): Fix an incorrect autocorrect for `Rails/Presence` when using `[]` method. ([@forresty][])
 * [#142](https://github.com/rubocop-hq/rubocop-rails/pull/142): Fix an incorrect autocorrect for `Rails/EnumHash` when using nested constants. ([@koic][])
+* [#136](https://github.com/rubocop-hq/rubocop-rails/pull/136): Fix a false positive for `Rails/ReversibleMigration` when using `change_default` with `:from` and `:to` options. ([@sinsoku][])
 
 ## 2.3.2 (2019-09-01)
 
@@ -93,3 +94,4 @@
 [@eugeneius]: https://github.com/eugeneius
 [@jas14]: https://github.com/jas14
 [@forresty]: https://github.com/forresty
+[@sinsoku]: https://github.com/sinsoku